# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SMART-ACADEMY API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlatformStats {
  totalStudents: number;
  totalCourses: number;
  totalInstructors: number;
  totalInstitutions: number;
  totalCountries: number;
  totalCertificates: number;
  totalJobPlacements: number;
  satisfactionRate: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  icon?: string | null;
  courseCount: number;
  /** @nullable */
  color?: string | null;
}

export interface Course {
  id: number;
  title: string;
  slug: string;
  description: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  instructorId?: number | null;
  /** @nullable */
  instructorName?: string | null;
  /** @nullable */
  instructorAvatar?: string | null;
  level: string;
  language: string;
  durationHours: number;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  rating: number;
  reviewCount: number;
  enrollmentCount: number;
  isFeatured: boolean;
  isFree: boolean;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}

export interface CourseInput {
  /** @minLength 1 */
  title: string;
  description: string;
  shortDescription?: string;
  thumbnail?: string;
  categoryId: number;
  instructorId?: number;
  level: string;
  language: string;
  durationHours: number;
  price: number;
  originalPrice?: number;
  isFeatured?: boolean;
  isFree?: boolean;
  tags?: string;
}

export interface CourseUpdate {
  title?: string;
  description?: string;
  shortDescription?: string;
  thumbnail?: string;
  price?: number;
  isFeatured?: boolean;
  level?: string;
  tags?: string;
}

export interface Enrollment {
  id: number;
  userId: number;
  courseId: number;
  /** @nullable */
  courseTitle?: string | null;
  /** @nullable */
  courseThumbnail?: string | null;
  progress: number;
  /** @nullable */
  completedAt?: string | null;
  enrolledAt: string;
  /** @nullable */
  certificateUrl?: string | null;
}

export interface EnrollmentInput {
  userId: number;
  courseId: number;
}

export interface ProgressUpdate {
  progress: number;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatar?: string | null;
  role: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  institution?: string | null;
  /** @nullable */
  country?: string | null;
  joinedAt: string;
  isVerified: boolean;
}

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
  avatar?: string;
  role: string;
  bio?: string;
  skills?: string;
  institution?: string;
  country?: string;
}

export interface StudentDashboard {
  userId: number;
  enrolledCourses: number;
  completedCourses: number;
  totalHoursLearned: number;
  certificates: number;
  currentStreak: number;
  skillBadges: number;
  recentEnrollments: Enrollment[];
  recommendedCourses: Course[];
}

export interface Event {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  thumbnail?: string | null;
  type: string;
  hostName: string;
  /** @nullable */
  hostAvatar?: string | null;
  scheduledAt: string;
  durationMinutes: number;
  registrationCount: number;
  /** @nullable */
  maxAttendees?: number | null;
  isLive: boolean;
  isFree: boolean;
  price: number;
  /** @nullable */
  meetingUrl?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}

export interface EventInput {
  /** @minLength 1 */
  title: string;
  description: string;
  thumbnail?: string;
  type: string;
  hostName: string;
  scheduledAt: string;
  durationMinutes: number;
  maxAttendees?: number;
  isFree?: boolean;
  price?: number;
  meetingUrl?: string;
  tags?: string;
}

export interface Job {
  id: number;
  title: string;
  company: string;
  /** @nullable */
  companyLogo?: string | null;
  location: string;
  type: string;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  currency: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  /** @nullable */
  skills?: string | null;
  experienceYears: number;
  category: string;
  applicationCount: number;
  isRemote: boolean;
  postedAt: string;
  /** @nullable */
  expiresAt?: string | null;
}

export interface JobInput {
  /** @minLength 1 */
  title: string;
  company: string;
  companyLogo?: string;
  location: string;
  type: string;
  salaryMin?: number;
  salaryMax?: number;
  currency?: string;
  description: string;
  requirements?: string;
  skills?: string;
  experienceYears?: number;
  category: string;
  isRemote?: boolean;
  expiresAt?: string;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  /** @nullable */
  thumbnail?: string | null;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  category: string;
  /** @nullable */
  tags?: string | null;
  viewCount: number;
  likeCount: number;
  readMinutes: number;
  publishedAt: string;
}

export interface BlogPostInput {
  /** @minLength 1 */
  title: string;
  excerpt?: string;
  content: string;
  thumbnail?: string;
  authorId: number;
  authorName: string;
  authorAvatar?: string;
  category: string;
  tags?: string;
  readMinutes?: number;
}

export interface CommunityPost {
  id: number;
  content: string;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  authorRole: string;
  /** @nullable */
  topic?: string | null;
  likeCount: number;
  replyCount: number;
  createdAt: string;
}

export interface CommunityPostInput {
  /** @minLength 1 */
  content: string;
  authorId: number;
  authorName: string;
  authorAvatar?: string;
  authorRole: string;
  topic?: string;
}

export interface SearchResults {
  query: string;
  type: string;
  courses: Course[];
  blogs: BlogPost[];
  jobs: Job[];
  events: Event[];
  totalResults: number;
}

export type ListCoursesParams = {
category?: string;
search?: string;
level?: string;
featured?: string;
limit?: number;
};

export type ListEnrollmentsParams = {
userId?: number;
};

export type ListUsersParams = {
role?: string;
};

export type ListEventsParams = {
upcoming?: string;
};

export type ListJobsParams = {
search?: string;
type?: string;
};

export type ListBlogPostsParams = {
category?: string;
limit?: number;
};

export type AiSearchParams = {
q: string;
type?: string;
};