# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SMART-ACADEMY API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/dashboard/{id}:
    get:
      operationId: getUserDashboard
      tags: [users]
      summary: Get student dashboard summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"

  # ── EVENTS ─────────────────────────────────────────────────────────────────
...
  /ai/search:
    get:
      operationId: aiSearch
      tags: [stats]
      summary: AI-powered smart search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
  /ai/recommendations/{userId}:
    get:
      operationId: getRecommendations
      tags: [stats]
      summary: Get AI recommendations for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        instructorId:
          type: ["integer", "null"]
        instructorName:
          type: ["string", "null"]
        instructorAvatar:
          type: ["string", "null"]
        level:
          type: string
        language:
          type: string
        durationHours:
          type: number
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        enrollmentCount:
          type: integer
        isFeatured:
          type: boolean
        isFree:
          type: boolean
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        [id, title, slug, description, categoryId, level, language, durationHours, price, rating, reviewCount, enrollmentCount, isFeatured, isFree, createdAt]
...
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        courseTitle:
          type: ["string", "null"]
        courseThumbnail:
          type: ["string", "null"]
        progress:
          type: integer
        completedAt:
          type: ["string", "null"]
        enrolledAt:
          type: string
        certificateUrl:
          type: ["string", "null"]
      required: [id, userId, courseId, progress, enrolledAt]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
        bio:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        institution:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        joinedAt:
          type: string
        isVerified:
          type: boolean
      required: [id, name, email, role, joinedAt, isVerified]
    UserInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        avatar:
          type: string
        role:
          type: string
        bio:
          type: string
        skills:
          type: string
        institution:
          type: string
        country:
          type: string
      required: [name, email, role]
    StudentDashboard:
      type: object
      properties:
        userId:
          type: integer
        enrolledCourses:
          type: integer
        completedCourses:
          type: integer
        totalHoursLearned:
          type: number
        certificates:
          type: integer
        currentStreak:
          type: integer
        skillBadges:
          type: integer
        recentEnrollments:
          type: array
          items:
            $ref: "#/components/schemas/Enrollment"
        recommendedCourses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
      required:
        [userId, enrolledCourses, completedCourses, totalHoursLearned, certificates, currentStreak, skillBadges, recentEnrollments, recommendedCourses]
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: ["string", "null"]
        type:
          type: string
        hostName:
          type: string
        hostAvatar:
          type: ["string", "null"]
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        registrationCount:
          type: integer
        maxAttendees:
          type: ["integer", "null"]
        isLive:
          type: boolean
        isFree:
          type: boolean
        price:
          type: number
        meetingUrl:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        [id, title, description, type, hostName, scheduledAt, durationMinutes, registrationCount, isLive, isFree, price, createdAt]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        companyLogo:
          type: ["string", "null"]
        location:
          type: string
        type:
          type: string
        salaryMin:
          type: ["number", "null"]
        salaryMax:
          type: ["number", "null"]
        currency:
          type: string
        description:
          type: string
        requirements:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        experienceYears:
          type: integer
        category:
          type: string
        applicationCount:
          type: integer
        isRemote:
          type: boolean
        postedAt:
          type: string
        expiresAt:
          type: ["string", "null"]
      required:
        [id, title, company, location, type, currency, description, experienceYears, category, applicationCount, isRemote, postedAt]
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: string
        thumbnail:
          type: ["string", "null"]
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        category:
          type: string
        tags:
          type: ["string", "null"]
        viewCount:
          type: integer
        likeCount:
          type: integer
        readMinutes:
          type: integer
        publishedAt:
          type: string
      required:
        [id, title, slug, content, authorId, authorName, category, viewCount, likeCount, readMinutes, publishedAt]
...
    SearchResults:
      type: object
      properties:
        query:
          type: string
        type:
          type: string
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        blogs:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        totalResults:
          type: integer
      required: [query, type, courses, blogs, jobs, events, totalResults]