# lib/api-zod/src/generated/types/studentDashboard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SMART-ACADEMY API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Course } from './course';
import type { Enrollment } from './enrollment';

export interface StudentDashboard {
  userId: number;
  enrolledCourses: number;
  completedCourses: number;
  totalHoursLearned: number;
  certificates: number;
  currentStreak: number;
  skillBadges: number;
  recentEnrollments: Enrollment[];
  recommendedCourses: Course[];
}