# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SMART-ACADEMY API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  avatar?: string | null;
  role: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  institution?: string | null;
  /** @nullable */
  country?: string | null;
  joinedAt: string;
  isVerified: boolean;
}